+++ CHANGELOG.rst
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
-v0.5.2 (unreleased)
+v0.5.2 (2019-11-25)
 ===================
 
 Bug fix release